Python-Django/project_management/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astor==0.8.1
asttokens==2.4.1
bcrypt==4.1.2
bleach==1.5.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.1
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.8
Django==4.1.13
django-admin-tools==0.9.3
django-cors-headers==4.3.1
django-environ==0.11.2
django-mongodb-engine==0.6.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
djangotoolbox==1.8.0
djongo==1.3.6
dnspython==2.5.0
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.13.1
fonttools==4.51.0
gast==0.5.4
google-api-core==2.19.1
google-auth==2.31.0
google-cloud-vision==3.7.2
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
html5lib==0.9999999
idna==3.7
imbalanced-learn==0.12.4
imblearn==0.0
iniconfig==2.0.0
ipykernel==6.29.2
ipython==8.21.0
jedi==0.19.1
joblib==1.4.0
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mongoengine==0.28.1
mysqlclient==2.2.5
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.5
packaging==23.2
pandas==2.2.2
parso==0.8.3
pbr==6.0.0
pillow==10.2.0
pipenv==2023.12.1
platformdirs==4.2.0
pluggy==1.5.0
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==4.25.2
psutil==5.9.8
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.17.2
PyJWT==1.7.1
pymongo==3.12.3
PyMySQL==1.1.0
pyparsing==3.1.2
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pyzmq==25.1.2
requests==2.32.3
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
setuptools==69.1.1
six==1.16.0
sqlparse==0.2.4
stack-data==0.6.3
stevedore==5.2.0
tensorboard==1.8.0
tensorflow @ https://storage.googleapis.com/tensorflow/mac/cpu/tensorflow-1.8.0-py3-none-any.whl#sha256=373c637f9c4f06346b1b02e280b30517d262a1651b252ff889f7c22716ef4548
termcolor==2.4.0
threadpoolctl==3.4.0
tornado==6.4
traitlets==5.14.1
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.25.1
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
virtualenvwrapper-win==1.2.7
wcwidth==0.2.13
Werkzeug==3.0.1
wheel==0.42.0
